DBZ-5071 Correctly handle NULL values in incremental snapshots
This makes changes in the Db2 connector for the corresponding commit in
the main Debezium project.  The main thing we have to do is override the
nullsSortLast function in Db2Connection to define how NULLs sort in Db2.
We also have to update the test suite, since the base test suite now
includes additional tests for working with keys that are nullable.